+++ src/libpcp/src/util.c
@@ -459,11 +459,11 @@ pmPrintValue(FILE *f,			/* output stream */
         break;
 
     case PM_TYPE_64:
-        fprintf(f, "%*lli", minwidth, a.ll);
+        fprintf(f, "%*lli", minwidth, (long long)a.ll);
         break;
 
     case PM_TYPE_U64:
-        fprintf(f, "%*llu", minwidth, a.ull);
+        fprintf(f, "%*llu", minwidth, (unsigned long long)a.ull);
         break;
 
     case PM_TYPE_FLOAT:
@@ -504,7 +504,7 @@ pmPrintValue(FILE *f,			/* output stream */
 	    if (val->value.pval->vlen == PM_VAL_HDR_SIZE + sizeof(__uint64_t)) {
 		__uint64_t	i;
 		memcpy((void *)&i, (void *)&val->value.pval->vbuf, sizeof(__uint64_t));
-		fprintf(f, "%*llu", minwidth, i);
+		fprintf(f, "%*llu", minwidth, (unsigned long long)i);
 		done = 1;
 	    }
 	    if (val->value.pval->vlen == PM_VAL_HDR_SIZE + sizeof(double)) {
