+++ src/pmnscomp/pmnscomp.c
@@ -599,7 +599,7 @@ Options:\n\
 		    printf("htab64[%3d]", j);
 		}
 		__ntohll((char *)&k64);	
-		printf(" %5lld", k64); 
+		printf(" %5lld", (long long)k64); 
 	    }
 	    printf("\n\n64-bit Node Table\n");
 	    for (j = 0; j < nodecnt; j++) {
@@ -615,7 +615,8 @@ Options:\n\
 	        __ntohll ((char *)&t.hash);
 
 		printf("node64[%4d] %6lld %6lld %6lld %6lld %-16.16s",
-		    j, t.parent, t.next, t.first, t.hash, _symbol+t.name);
+		    j, (long long)t.parent, (long long)t.next,
+		    (long long)t.first, (long long)t.hash, _symbol+t.name);
 		if (t.first == -1) {
 		    printf(" %s", pmIDStr(__htonpmID(_nodetab64[j].pmid)));
 		}
